# Arm backend: Use bucket approach in fuse_equal_placeholder_pass