Verified to catch same dupes as before in lstm and mv2.

Instead of comparing all placeholders to each other, compute a hash use as key in dictionary. Equal placeholder -> equal key. If an entry in the dictionary has multiple values, we have duplicates.

This is a ~O(N) algorithm compared to earlier  O(N^2). This can be seen by measuring the speedup for lstm vs. mv2

lstm: 120 placeholders (116 dupes) 0.4s -> 0.3s

mv2: 318 placeholders (98 dupes) ~15s -> 0.5s

cc @digantdesai @freddan80 @per @zingo @oscarandersson8218